cleanup/make components
