+++ apps/main/app/hooks/useTierData.ts
@@ -1,5 +1,5 @@
 /**
- * Custom hook for fetching and formatting tier data with theme colors
+ * Custom hook for fetching and formatting tier data (with theme colors)
  */
 
 import { useMemo } from 'react'
@@ -118,7 +118,7 @@ export function useScenarioTiers(scenarioId: string | null) {
       "Salmon abundance",
     ]
     
-    // Create a map for quick lookup
+    // Quick lookup
     const tiersByDisplayName = new Map()
     allTiers.forEach((tier) => {
       const displayName = mapShortCodeToDisplayName(tier.short_code, tierMapping)
@@ -154,6 +154,151 @@ export function useScenarioTiers(scenarioId: string | null) {
     outcomeNames,
     isLoading: scenarioLoading || tiersLoading || mappingLoading,
     error: scenarioError || tiersError || mappingError
+  }
+}
+
+// Hook for fetching multiple scenarios
+export function useMultipleScenarioTiers() {
+  const theme = useTheme()
+  
+  // Fetch scenarios in parallel (React hooks must be at top level, not in callbacks like map())
+  const s0020Result = useSWR(`/api/tiers/scenarios/s0020/tiers`, () => fetchScenarioTiers("s0020"))
+  const s0021Result = useSWR(`/api/tiers/scenarios/s0021/tiers`, () => fetchScenarioTiers("s0021"))
+  const s0011Result = useSWR(`/api/tiers/scenarios/s0011/tiers`, () => fetchScenarioTiers("s0011"))
+
+
+  const { data: allTiers, error: tiersError, isLoading: tiersLoading } = useSWR(
+    '/api/tiers/list',
+    fetchTierList
+  )
+
+  const { data: tierMapping, error: mappingError, isLoading: mappingLoading } = useSWR(
+    '/api/tiers/mapping',
+    getTierMapping
+  )
+
+  // Convert all scenario data to chart format
+  const allChartData = useMemo(() => {
+    if (!tierMapping) return {}
+
+    const themeColors = {
+      tier1: theme.palette.tiers.tier1,
+      tier2: theme.palette.tiers.tier2,
+      tier3: theme.palette.tiers.tier3,
+      tier4: theme.palette.tiers.tier4,
+    }
+
+    const result: Record<string, Record<string, Array<{
+      label: string
+      color: string
+      value: number
+    }>>> = {}
+
+    // s0020
+    if (s0020Result.data) {
+      const converted: Record<string, Array<{ label: string; color: string; value: number }>> = {}
+      Object.entries(s0020Result.data.tiers).forEach(([shortCode, tierInfo]: [string, any]) => {
+        const displayName = mapShortCodeToDisplayName(shortCode, tierMapping)
+        if (tierInfo.type === "multi_value" && tierInfo.data) {
+          converted[displayName] = convertMultiValueToChartData(
+            { name: tierInfo.name, type: "multi_value", data: tierInfo.data, total: tierInfo.total || 0 },
+            themeColors
+          )
+        } else if (tierInfo.type === "single_value" && tierInfo.level) {
+          converted[displayName] = convertSingleValueToChartData(tierInfo.level, themeColors)
+        }
+      })
+      result["s0020"] = converted
+    }
+
+    // s0021
+    if (s0021Result.data) {
+      const converted: Record<string, Array<{ label: string; color: string; value: number }>> = {}
+      Object.entries(s0021Result.data.tiers).forEach(([shortCode, tierInfo]: [string, any]) => {
+        const displayName = mapShortCodeToDisplayName(shortCode, tierMapping)
+        if (tierInfo.type === "multi_value" && tierInfo.data) {
+          converted[displayName] = convertMultiValueToChartData(
+            { name: tierInfo.name, type: "multi_value", data: tierInfo.data, total: tierInfo.total || 0 },
+            themeColors
+          )
+        } else if (tierInfo.type === "single_value" && tierInfo.level) {
+          converted[displayName] = convertSingleValueToChartData(tierInfo.level, themeColors)
+        }
+      })
+      result["s0021"] = converted
+    }
+
+    // s0011
+    if (s0011Result.data) {
+      const converted: Record<string, Array<{ label: string; color: string; value: number }>> = {}
+      Object.entries(s0011Result.data.tiers).forEach(([shortCode, tierInfo]: [string, any]) => {
+        const displayName = mapShortCodeToDisplayName(shortCode, tierMapping)
+        if (tierInfo.type === "multi_value" && tierInfo.data) {
+          converted[displayName] = convertMultiValueToChartData(
+            { name: tierInfo.name, type: "multi_value", data: tierInfo.data, total: tierInfo.total || 0 },
+            themeColors
+          )
+        } else if (tierInfo.type === "single_value" && tierInfo.level) {
+          converted[displayName] = convertSingleValueToChartData(tierInfo.level, themeColors)
+        }
+      })
+      result["s0011"] = converted
+    }
+
+    return result
+  }, [s0020Result.data, s0021Result.data, s0011Result.data, tierMapping, theme.palette.tiers])
+
+  // Get outcome names from first scenario (structure should be same)
+  const outcomeNames = useMemo(() => {
+    if (!allTiers || !tierMapping) return []
+    
+    const desiredOrder = [
+      "Community deliveries",
+      "Agricultural revenue", 
+      "Environmental flows",
+      "Delta ecology",
+      "Freshwater for Delta exports",
+      "Freshwater for in-Delta uses",
+      "Reservoir storage",
+      "Groundwater storage",
+      "Salmon abundance",
+    ]
+    
+    const tiersByDisplayName = new Map()
+    allTiers.forEach((tier) => {
+      const displayName = mapShortCodeToDisplayName(tier.short_code, tierMapping)
+      tiersByDisplayName.set(displayName, tier)
+    })
+    
+    const firstScenarioData = s0020Result.data
+    return desiredOrder.map((displayName) => {
+      const tier = tiersByDisplayName.get(displayName)
+      if (!tier) {
+        return {
+          shortCode: 'MISSING',
+          name: displayName,
+          displayName,
+          isActive: false
+        }
+      }
+      
+      return {
+        shortCode: tier.short_code,
+        name: tier.name,
+        displayName,
+        isActive: firstScenarioData?.tiers[tier.short_code] !== undefined
+      }
+    })
+  }, [allTiers, tierMapping, s0020Result.data])
+
+  const isLoading = s0020Result.isLoading || s0021Result.isLoading || s0011Result.isLoading || tiersLoading || mappingLoading
+  const error = s0020Result.error || s0021Result.error || s0011Result.error || tiersError || mappingError
+
+  return {
+    allChartData,
+    outcomeNames,
+    isLoading,
+    error
   }
 }
